refactored evaluation
