+++ bundesliga/src/bundesliga/pipelines/data_science/__init__.py
@@ -3,8 +3,20 @@ This is a boilerplate pipeline 'data_science'
 generated using Kedro 0.19.10
 """
 
-from .pipeline import create_pipeline
+from .pipeline import (
+    create_model_definition_node,
+    create_subpipeline_for_match,
+    create_walk_forward_pipeline,
+    eval_pipeline,
+    ml_pipeline,
+)
 
-__all__ = ["create_pipeline"]
+__all__ = [
+    "ml_pipeline",
+    "create_model_definition_node",
+    "eval_pipeline",
+    "create_walk_forward_pipeline",
+    "create_subpipeline_for_match",
+]
 
 __version__ = "0.1"
